backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.13.0
scikit-learn==1.3.2
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
gunicorn==21.2.0
psycopg2-binary==2.9.9